Rename: interferometer_monitor.py => em27_monitor.py
This way the file name is consistent with the class name.
