+++ finesse/gui/em27_monitor.py
