+++ finesse/gui/main_window.py
@@ -5,7 +5,7 @@ from PySide6.QtWidgets import QGridLayout, QGroupBox, QHBoxLayout, QMainWindow, 
 
 from ..config import APP_NAME
 from .data_file_view import DataFileControl
-from .interferometer_monitor import EM27Monitor
+from .em27_monitor import EM27Monitor
 from .measure_script.script_view import ScriptControl
 from .opus_view import OPUSControl
 from .serial_view import SerialPortControl
